requirements.txt
psycopg2-binary==2.9.2
pyfiglet==0.8.post1
termcolor==1.1.0
simple-colors==0.1.5